feat: implement deleting elements from favourites when they are deleted
