+++ src/favs/services/fav.service.ts
@@ -4,7 +4,7 @@ import { IAlbum } from "../../albums/interfaces/album.interface";
 import { ITrack } from "../../tracks/interfaces/track.interface";
 import { IFavBody } from "../interfaces/fav-body.interface";
 import { IFav } from "../interfaces/fav.interface";
-import { request } from "http";
+import { sendRequest } from "../../sendRequest";
 import { validate } from "uuid"
 
 const checkValidation = (id) => { 
@@ -35,15 +35,15 @@ export class FavService {
             tracks: [],
         };
         for (let id of this.fav.artists) {
-            const answer = await this.sendRequest(`http://${host}/artist/${id}`) as IArtist;
+            const answer = await sendRequest(`http://${host}/artist/${id}`) as IArtist;
             favBody.artists.push(answer);
         }
         for (let id of this.fav.albums) {
-            const answer = await this.sendRequest(`http://${host}/album/${id}`) as IAlbum;
+            const answer = await sendRequest(`http://${host}/album/${id}`) as IAlbum;
             favBody.albums.push(answer);
         }
         for (let id of this.fav.tracks) {
-            const answer = await this.sendRequest(`http://${host}/track/${id}`) as ITrack;
+            const answer = await sendRequest(`http://${host}/track/${id}`) as ITrack;
             favBody.tracks.push(answer);
         }
         return favBody;
@@ -51,7 +51,7 @@ export class FavService {
 
     async addArtist(id: string, host: string) {
         checkValidation(id);
-        const answer = await this.sendRequest(`http://${host}/artist/${id}`) as IAnswer;
+        const answer = await sendRequest(`http://${host}/artist/${id}`) as IAnswer;
         if("statusCode" in answer) {
             throw new UnprocessableEntityException(`Artist with id: "${id}" didn't exist`);
         } else {
@@ -73,7 +73,7 @@ export class FavService {
 
     async addAlbum(id: string, host: string) {
         checkValidation(id);
-        const answer = await this.sendRequest(`http://${host}/album/${id}`) as IAnswer;
+        const answer = await sendRequest(`http://${host}/album/${id}`) as IAnswer;
         if("statusCode" in answer) {
             throw new UnprocessableEntityException(`Artist with id: "${id}" didn't exist`);
         } else {
@@ -95,7 +95,7 @@ export class FavService {
 
     async addTrack(id: string, host: string) {
         checkValidation(id);
-        const answer = await this.sendRequest(`http://${host}/track/${id}`) as IAnswer;
+        const answer = await sendRequest(`http://${host}/track/${id}`) as IAnswer;
         if("statusCode" in answer) {
             throw new UnprocessableEntityException(`Artist with id: "${id}" didn't exist`);
         } else {
@@ -113,43 +113,5 @@ export class FavService {
             this.fav.tracks.splice(trackIndex, 1);
             return `Track id: "${id}" successfully deleted from favourites`;
         }
-    }
-
-    async sendRequest(path, method="GET", body={}, headers = {}) {
-        return new Promise((resolve, reject) => {
-            const postData = JSON.stringify(body);
-            let length = 0;
-            if(method !== "GET" && method !== "DELETE") {
-                length = Buffer.byteLength(postData);
-            }
-            const options = {
-                method: method,
-                headers: {
-                  'Content-Type': 'application/json',
-                  'Content-Length': length
-                }
-            };
-            const req = request(path, options, (res) => {
-                let body = "";
-                res.setEncoding('utf8');
-                res.on('data', (chunk) => {
-                  body += chunk;
-                });
-                res.on('end', () => {
-                    resolve(JSON.parse(body));
-                });
-            });
-            for (let [key, value] of Object.entries(headers)) {
-                req.setHeader(key, value as string);
-            }
-            req.on('error', (e) => {
-                console.error(`problem with request: ${e.message}`);
-                reject(e);
-            });
-            if(method !== "GET" && method !== "DELETE") {
-                req.write(postData);
-            }
-            req.end();
-        });
     }
 }
